requirements.txt
Flask-Cors>=3.0.10
openai>=0.27.2
pinecone-client>=2.0.13
PyPDF2>=2.10.4
numpy>=1.23.2
scikit-learn>=1.1.2
docx2txt>=0.8
flask>=1.1.4
jinja2>=3.0.1
PyYAML>=6.0
tiktoken>=0.1.2
openpyxl
pandas
gunicorn
beautifulsoup4